% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_maverick_con.R
\name{aggregate_maverick_con}
\alias{aggregate_maverick_con}
\title{Load the most-conservative aggregation of MAVERICK}
\usage{
aggregate_maverick_con(data)
}
\arguments{
\item{data}{The MAVERICK event report level dataset. Already pre-loaded.}
}
\value{
Returns a dataframe of the most-conservative aggregation of the MAVERICK dataset.
}
\description{
This convenience function aggregates the MAVERICK event report data to the event level using the most-conservative aggregation model.
}
\examples{
maverick_conservative <- aggregate_maverick_con()
}
