% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_maverick_inf.R
\name{aggregate_maverick_inf}
\alias{aggregate_maverick_inf}
\title{Load the most-informative aggregation of MAVERICK}
\usage{
aggregate_maverick_inf(data)
}
\arguments{
\item{data}{The MAVERICK event report level dataset. Already pre-loaded.}
}
\value{
Returns a dataframe of the most-informative aggregation of the MAVERICK dataset.
}
\description{
This convenience function aggregates the MAVERICK event report data to the event level using the most-informative aggregation model.
}
\examples{
\donttest{
maverick_informative <- aggregate_maverick_inf()
}
}
