% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_maverick_rep.R
\name{aggregate_maverick_rep}
\alias{aggregate_maverick_rep}
\title{Load the most-representative aggregation of MAVERICK}
\usage{
aggregate_maverick_rep(data)
}
\arguments{
\item{data}{The MAVERICK event report level dataset. Already pre-loaded.}
}
\value{
Returns a dataframe of the most-representative aggregation of the MAVERICK dataset
}
\description{
This convenience function aggregates the MAVERICK event report data to the event level using the most-representative aggregation model.
}
\examples{
\donttest{
maverick_representative <- aggregate_maverick_rep()
}
}
