% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_maverick_typ.R
\name{aggregate_maverick_typ}
\alias{aggregate_maverick_typ}
\title{Load the most-typical aggregation of MAVERICK}
\usage{
aggregate_maverick_typ(data)
}
\arguments{
\item{data}{The MAVERICK event report level dataset. Already pre-loaded.}
}
\value{
Returns a dataframe of the most-typical aggregation of MAVERICK.
}
\description{
This convenience function aggregates the MAVERICK event report data to the event level using the most-typical aggregation model.
}
\examples{
maverick_typical <- aggregate_maverick_typ()
}
