% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_strings.R
\name{aggregate_strings}
\alias{aggregate_strings}
\title{A function for combining strings from a character variable}
\usage{
aggregate_strings(str_var)
}
\arguments{
\item{str_var}{A character vector.}
}
\value{
Returns a single character string with unique strings concatenated by semicolons.
}
\description{
A function for combining strings from a character variable
}
\examples{
aggregate_strings(c("apple", "banana", "apple", "Unknown", "orange", " "))
}
