% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode_binary.R
\name{calc_mode_binary}
\alias{calc_mode_binary}
\title{Calculate mode of a binary numeric vector}
\usage{
calc_mode_binary(x)
}
\arguments{
\item{x}{A numeric vector consisting only of binary values (0 and 1).}
}
\value{
Returns a numeric vector representing the mode value. Returns 1 if there is a tie. Returns `NA` if the vector is empty.
}
\description{
Calculate mode of a binary numeric vector
}
\examples{
calc_mode_binary(c(0, 1, 1, 0, 1))
}
