% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode_date.R
\name{calc_mode_date}
\alias{calc_mode_date}
\title{Calculate mode of date vector}
\usage{
calc_mode_date(x)
}
\arguments{
\item{x}{A character vector where each element is a date in "YYYY-MM-DD" format.}
}
\value{
Returns a date vector representing the modal date, or the mean of the modal dates if there is a tie.
}
\description{
Calculate mode of date vector
}
\examples{
calc_mode_date(c("2021-01-01", "2021-01-02", "2021-01-01"))
}
