% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode_numeric.R
\name{calc_mode_numeric}
\alias{calc_mode_numeric}
\title{Calculate mode of numeric vector}
\usage{
calc_mode_numeric(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns a numeric vector representing the mode value. Returns the smallest mode value if multiple modes exist, and NA if the vector is empty or contains non-numeric elements.
}
\description{
This function calculates the mode of a given numeric vector, and returns the smallest mode value if multiple modes exist.
}
\examples{
calc_mode_numeric(c(1, 2, 2, 3, 4, 4))
}
