% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventreport-package.R
\docType{package}
\name{eventreport}
\alias{eventreport-package}
\alias{eventreport}
\title{eventreport: Tools for Aggregating Event Data}
\description{
The `eventreport` package provides tools for analyzing and aggregating event
data on political violence. It includes functions for processing, visualizing,
and summarizing event data.
}
\section{Main Functions}{

- `aggregate_maverick_con`: Function for [describe what this does].
- `mean_dscore`: Function for [describe what this does].
}

\references{
Sebastian van Baalen & Kristine Hoglund (2024) Trials and Triangulations:
Aggregation Problems in Event Data on Political Violence. Uppsala: Uppsala University.
}
\author{
Sebastian van Baalen & Kristine Hoglund (2024)
}
\keyword{internal}
