% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_dscore.R
\name{mean_dscore}
\alias{mean_dscore}
\title{A function for calculating mean divergence scores}
\usage{
mean_dscore(data, group_var, variables, normalize = FALSE, plot = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{group_var}{A string indicating the grouping variable in the data frame.}

\item{variables}{A character vector of column names for which the divergence score will be calculated.}

\item{normalize}{Logical, indicating whether to normalize the scores by the total number of unique values across the dataset.}

\item{plot}{Logical, indicating whether to return a ggplot object visualizing the scores instead of a data frame.}
}
\value{
A data frame or a ggplot object, depending on the value of the \code{plot} argument.
If \code{plot} is FALSE, returns a data frame with variables and their corresponding scores.
If \code{plot} is TRUE, returns a ggplot object displaying the scores.
}
\description{
This function computes the mean divergence of the number of unique values minus one for each specified variable within each group. Optionally, it normalizes this divergence by the total number of unique values that the variable takes in the dataset. Users can also use the function to directly plot the output as a ggplot object.
}
\examples{
df <- data.frame(
  event_id = c(1, 1, 2, 2, 3),
  country = c("US", "US", "UK", "UK", "CA"),
  actor1 = c("Actor A", "Actor B", "Actor B", "Actor C", "Actor D"),
  deaths_best = c(10, 20, 5, 15, 10)
)
mean_dscore(df, "event_id", c("country", "actor1", "deaths_best"), normalize = TRUE, plot = TRUE)
}
